requirements.txt
anthropic==0.96.0
python-dotenv==1.2.2
requests==2.33.1
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
gunicorn==23.0.0
supabase==2.8.0
